Add src/utils/mainUtils.ts
